use atlas.config.yml
